+++ cirq-core/cirq/protocols/json_test_data/ReadoutExperimentResult.json
@@ -1,0 +1,23 @@
+{
+  "cirq_type": "ReadoutExperimentResult",
+  "zero_state_errors": [
+    [
+      {
+        "cirq_type": "LineQubit",
+        "x": 0
+      },
+      0.1
+    ]
+  ],
+  "one_state_errors": [
+    [
+      {
+        "cirq_type": "LineQubit",
+        "x": 0
+      },
+      0.2
+    ]
+  ],
+  "repetitions": 1000,
+  "timestamp": 0.3
+}
