+++ cirq-core/cirq/protocols/json_test_data/ReadoutExperimentResult.repr
@@ -1,0 +1,1 @@
+cirq.experiments.ReadoutExperimentResult(zero_state_errors={cirq.LineQubit(0): 0.1}, one_state_errors={cirq.LineQubit(0): 0.2}, repetitions=1000, timestamp=0.3)
