+++ cirq-core/cirq/protocols/json_test_data/SingleQubitReadoutCalibrationResult.json_inward
